+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Changed
 
 - 📦️(front) update react
+- ✨(chat) Generate and edit conversation title
 
 ### Fixed